made a lot of mini edits

+++ src/main.py
@@ -15,13 +15,14 @@ License:
     See the License for the specific language governing permissions and
     limitations under the License.
 
+Foreword:
+    I'm trying to observe standards of code writing on Python(PEP8)
+
 Usage:
     python ./src/main.py or python3 ./src/main.py
 
 Functions:
-    create_img_folder_if_not_exist
-    delete_all_images
-    somethings_went_wrong
+    do_delete_all_images
     main
 """
 
@@ -30,50 +31,44 @@ import os
 
 from core import ScrabImgFromLightShot
 
-# Check file img/temp
-def create_img_folder_if_not_exist():
-    """Create folder img/ if it doesn't exist.
 
-    If folder img/ doesn't exist will be an error, so need to check this every
-    time when program starts.
-
-    Returns:
-        None
-    """
-    if not os.path.exists("img/"):
-        os.makedirs(os.path.dirname("img/"))
-
-
-def delete_all_images():
+def do_delete_all_images():
     """Delete all images from folder img/
 
-    It did for convince, to don't delete it manually. And this function call
-    every time when program starts.
+    It did it for convince, to don't delete it manually. There are 2 choices to
+    delete it, or not, for convince.
 
     Returns:
         None
     """
-    # Need to do accepting, "Do you want to delete all uploaded images?"
-    # If yes - delete, if no - quo it
-    for folder, _, files in os.walk("img/"):
-        for file in files:
-            os.remove(folder + file)
+    if len(os.listdir("img/")) != 0:
+        permission_for_delete_images = input(
+            "Do you want delete all images from folder, (y - yes, n - no) "
+        )
+        if permission_for_delete_images == "y":
+            for folder, _, files in os.walk("img/"):
+                for file in files:
+                    os.remove(folder + file)
+        elif permission_for_delete_images == "n":
+            pass
+        else:
+            print("Incorrect input, try again")
 
 
 def main(number_of_images, debug_mod=False):
     """Enter point in program.
 
     Arguments:
-        number_of_images (int):
-        debug_mod (bool):
+        number_of_images (int): number of images, that need to download
+        debug_mod (bool): turn on/off debug mod
 
     Returns:
         None
     """
-    create_img_folder_if_not_exist()
-    delete_all_images()
+    do_delete_all_images()
 
     start_program_time = time.time()
+    # Need to rewrite this comment, it doesn't give enough truly information
     # When this code(in while loop) executing, sometimes errors occur due to
     # the lack of a link, or it could not be obtained, and so that the program
     # does not stop - an endless loop wrapped in try/except, which absorbs all
@@ -83,13 +78,26 @@ def main(number_of_images, debug_mod=False):
     while number_of_images != len(os.listdir("img/")):
         try:
             ScrabImgFromLightShot.download_img()
-        except Exception as err:
+        except Exception as error:
             if debug_mod:
-                print("SOMETHING WENT WRONG!")
-                print(err)
+                print("SOMETHING WENT WRONG!:")
+                print(error)
             pass
+
     print(f"All time: {int(time.time() - start_program_time)}")
 
 
 if __name__ == "__main__":
-    main(int(input("How many images to download(default=10): ") or 10))
+    number_of_images = int(input("How many images to download(default=10): ")
+                           or 10)
+    # May occur, that folder img/ doesn't exist, and will be an error.
+    try:
+        main(int(number_of_images))
+    except FileNotFoundError:
+        # If you work in pycharm or another code editor/IDE, check the
+        # "working directory"(in Configuration, where Running the scripts)
+        # there must be root name of this project, not folder src or else.
+        # Folder img must be in root directory.
+        os.mkdir("img")
+        main(int(number_of_images))
+
